cambio las variables int por byte para hacer un mejor uso de la memoria

+++ despertador_matriz_cuatro.ino
@@ -12,10 +12,10 @@ class Reloj {
 public:
   char horaMostrar[5];
   char alarmaMostrar[5];
-  int hora;
-  int minutos;
-  int horaAlarma;
-  int minutosAlarma;
+  byte hora;
+  byte minutos;
+  byte horaAlarma;
+  byte minutosAlarma;
   bool sonar;
 
   Reloj() {
@@ -46,39 +46,39 @@ public:
     return alarmaMostrar;
   }
 
-  void setHora(int h) {
+  void setHora(byte h) {
     hora = h;
     setDisplay();
   }
 
-  int getHora() {
+  byte getHora() {
     return hora;
   }
 
-  void setHoraAlarma(int h) {
+  void setHoraAlarma(byte h) {
     horaAlarma = h;
     setDisplayAlarma();
   }
 
-  int getHoraAlarma() {
+  byte getHoraAlarma() {
     return horaAlarma;
   }
 
-  void setMinutos(int m) {
+  void setMinutos(byte m) {
     minutos = m;
     setDisplay();
   }
 
-  int getMinutos() {
+  byte getMinutos() {
     return minutos;
   }
 
-  void setMinutosAlarma(int m) {
+  void setMinutosAlarma(byte m) {
     minutosAlarma = m;
     setDisplayAlarma();
   }
 
-  int getMinutosAlarma() {
+  byte getMinutosAlarma() {
     return minutosAlarma;
   }
 
@@ -91,31 +91,31 @@ public:
   }
 };
 
-const int tiempoRebote = 180;
+const byte  tiempoRebote = 180;
 unsigned long ultimaPresionBtn[] = { 0, 0, 0, 0, 0 };
 unsigned long currentMillis = millis();
 
 Reloj miReloj;
 RTC_DS3231 rtc;
 
-int alarmaMin = 0;
-int alarmaHora = 0;
+byte  alarmaMin = 0;
+byte  alarmaHora = 0;
 
-const int pinCS = 10;
-const int numberOfHorizontalDisplays = 4;
-const int numberOfVerticalDisplays = 1;
+const byte  pinCS = 10;
+const byte  numberOfHorizontalDisplays = 4;
+const byte  numberOfVerticalDisplays = 1;
 Max72xxPanel matrix = Max72xxPanel(pinCS, numberOfHorizontalDisplays, numberOfVerticalDisplays);
 
-const int buzzer = 9;
-const int button0 = 3;
-const int button1 = 4;
-const int button2 = 5;
-const int button3 = 6;
-const int button4 = 2;
+const byte  buzzer = 9;
+const byte  button0 = 3;
+const byte  button1 = 4;
+const byte  button2 = 5;
+const byte  button3 = 6;
+const byte  button4 = 2;
 unsigned long tiempoInicio = millis();
 unsigned long esperaEntreSonidos = 1000;
-int modo = 0;
-int brillo = 0;
+byte modo = 0;
+byte brillo = 0;
 char caracter = 'E';
 char diasDeLaSemana[] = { 'D', 'L', 'M', 'M', 'J', 'V', 'S' };
 bool configInicial = true;
@@ -123,9 +123,9 @@ bool dolorDeCabeza = false;
 bool vueltaUnica = true;
 bool ultimaConfigAlarma = false;
 bool visualizacionSegundos = false;
-int contadorSegundos = 0;
+byte contadorSegundos = 0;
 bool entraPrimeraVez = true;
-int caminoSegundo = 3;
+byte caminoSegundo = 3;
 bool incrementando = true;
 bool visualizacionSegundosAnterior = false;
 
@@ -262,7 +262,7 @@ void pantallaHora() {
   pilotoDelSegundo();
   
   if (dolorDeCabeza) {
-    matrix.drawPixel(8, 7, HIGH);
+    matrix.drawPixel(8, 7, HIGH);    
   }
 
   matrix.write();
